Fix incorrect return code from local file checks
The code for checking for and sourcing local variants of .aliases,
.bash_profile, .bashrc, .exports, and .functions returns an error code
by default if the local file doesn't exist. This is buggy since this is
an optional check, and running/sourcing any of these scripts in a
script with set -e enabled causes the script to fail due to this.